New API for the NUTS kernel
