+++ blackjax/nuts_base.py
@@ -10,11 +10,10 @@ import blackjax.inference.hmc.metrics as metrics
 import blackjax.inference.hmc.proposal as proposal
 import blackjax.inference.hmc.termination as termination
 import blackjax.inference.hmc.trajectory as trajectory
-from blackjax.base import SamplingAlgorithm
-from blackjax.hmc_base import HMCState, hmc_init
+from blackjax.hmc_base import HMCState
 from blackjax.types import Array, PRNGKey, PyTree
 
-__all__ = ["NUTSInfo", "nuts", "nuts_kernel"]
+__all__ = ["NUTSInfo", "nuts_kernel"]
 
 
 class NUTSInfo(NamedTuple):
@@ -54,30 +53,6 @@ class NUTSInfo(NamedTuple):
     num_trajectory_expansions: int
     integration_steps: int
     acceptance_probability: float
-
-
-def nuts(
-    logprob_fn: Callable,
-    step_size: float,
-    inverse_mass_matrix: Array,
-    *,
-    max_num_doublings: int = 10,
-    integrator: Callable = integrators.velocity_verlet,
-    divergence_threshold: int = 1000,
-) -> SamplingAlgorithm:
-
-    kernel = nuts_kernel(integrator, divergence_threshold, max_num_doublings)
-
-    def init_fn(position: PyTree):
-        return jax.jit(hmc_init, static_argnums=(1,))(position, logprob_fn)
-
-    def step_fn(rng_key: PRNGKey, state: HMCState) -> Tuple[HMCState, NUTSInfo]:
-        # `np.ndarray` and `DeviceArray`s are not hashable and thus cannot be used as static arguments.`
-        # Workaround: https://github.com/google/jax/issues/4572#issuecomment-709809897
-        kernel_fn = jax.jit(kernel, static_argnames=["logprob_fn", "step_size"])
-        return kernel_fn(rng_key, state, logprob_fn, step_size, inverse_mass_matrix)
-
-    return SamplingAlgorithm(init_fn, step_fn)
 
 
 def nuts_kernel(
